requirements.txt
boto3
flask
numpy
pydicom
pyyaml
requests-toolbelt
SimpleITK
tensorflow==1.15
opencv-python
Pillow